structure organization
- Put Task into its own file, and created TaskViewModel to handle all functions of a Task object.
- DateFormattedView is now its own file.
