+++ To-Do/To_DoApp.swift
@@ -11,7 +11,7 @@ import SwiftUI
 struct To_DoApp: App {
     var body: some Scene {
         WindowGroup {
-            ContentView()
+            Home()
         }
     }
 }
